GenerateExcelReport/requirements.txt
azure-functions
azure-identity
azure-storage-blob
azure-kusto-data
azure-kusto-ingest
pandas
openpyxl
